+++ src/ModularPipelines.Azure/Extensions/AzureExtensions.cs
@@ -43,5 +43,5 @@ public static class AzureExtensions
         return services.AddSingleton(new ArmClient(tokenCredential));
     }
 
-    public static IAzure Azure(this IPipelineContext context) => context.ServiceProvider.GetRequiredService<IAzure>();
+    public static IAzure Azure(this IPipelineHookContext context) => context.ServiceProvider.GetRequiredService<IAzure>();
 }