+++ src/ModularPipelines.Build/Attributes/SkipOnMainBranch.cs
@@ -7,7 +7,7 @@ namespace ModularPipelines.Build.Attributes;
 public class SkipOnMainBranch : MandatoryRunConditionAttribute
 {
     /// <inheritdoc/>
-    public override Task<bool> Condition(IPipelineContext pipelineContext)
+    public override Task<bool> Condition(IPipelineHookContext pipelineContext)
     {
         return Task.FromResult(pipelineContext.Git().Information.BranchName != "main");
     }