+++ src/ModularPipelines.Git/Attributes/RunOnlyIfBranchStartsWithAttribute.cs
@@ -15,7 +15,7 @@ public class RunOnlyIfBranchStartsWithAttribute : MandatoryRunConditionAttribute
         BranchNamePrefix = branchNamePrefix;
     }
     
-    public override Task<bool> Condition(IPipelineContext pipelineContext)
+    public override Task<bool> Condition(IPipelineHookContext pipelineContext)
     {
         return Task.FromResult(pipelineContext.Git().Information.BranchName?.StartsWith(BranchNamePrefix) ?? false);
     }