+++ src/ModularPipelines.TeamCity/Extensions/TeamCityExtensions.cs
@@ -23,5 +23,5 @@ public static class TeamCityExtensions
         return services;
     }
 
-    public static ITeamCity TeamCity(this IPipelineContext context) => context.ServiceProvider.GetRequiredService<ITeamCity>();
+    public static ITeamCity TeamCity(this IPipelineHookContext context) => context.ServiceProvider.GetRequiredService<ITeamCity>();
 }