+++ src/ModularPipelines.Terraform/Extensions/TerraformExtensions.cs
@@ -23,5 +23,5 @@ public static class TerraformExtensions
         return services;
     }
 
-    public static ITerraform Terraform(this IPipelineContext context) => context.ServiceProvider.GetRequiredService<ITerraform>();
+    public static ITerraform Terraform(this IPipelineHookContext context) => context.ServiceProvider.GetRequiredService<ITerraform>();
 }