+++ src/ModularPipelines/Attributes/RunOnLinuxAttribute.cs
@@ -7,7 +7,7 @@ namespace ModularPipelines.Attributes;
 public class RunOnLinuxAttribute : RunConditionAttribute
 {
     /// <inheritdoc/>
-    public override Task<bool> Condition(IPipelineContext pipelineContext)
+    public override Task<bool> Condition(IPipelineHookContext pipelineContext)
     {
         return Task.FromResult(OperatingSystem.IsLinux());
     }