+++ test/ModularPipelines.UnitTests/ModuleHistoryTests.cs
@@ -21,7 +21,7 @@ public class ModuleHistoryTests
 
     private class SkipRunConditionAttribute : RunConditionAttribute
     {
-        public override Task<bool> Condition(IPipelineContext pipelineContext)
+        public override Task<bool> Condition(IPipelineHookContext pipelineContext)
         {
             return false.AsTask();
         }