+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,2 @@
-*   Breaking: Added `IPipelineContext` parameter into `IModuleResultRepository`
+*   Breaking: Added `IPipelineHookContext` parameter into `IModuleResultRepository`
+*   Breaking: Hooks now accept a `IPipelineHookContext` rather than a `IPipelineContext`